pyproject.toml
[tool.poetry]
name = "rag-langchain-llm-chatbot"
version = "0.1.0"
description = ""
authors = ["Victor Ortiz"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
asyncio = "^3.4.3"
uuid = "^1.30"
langsmith = "^0.1.83"
fastapi = "^0.111.0"
langserve = "^0.2.2"
pydantic = "^2.8.2"
uvicorn = "^0.30.1"
logging = "^0.4.9.6"
bs4 = "^0.0.2"
langchain-openai = "^0.1.14"
langchain = "^0.2.6"
langchain-community = "^0.2.6"
langchain-core = "^0.2.11"
weaviate-client = "3.26.5"
langchain-anthropic = "^0.1.19"
langchain-fireworks = "^0.1.4"
langchain-google-genai = "^1.0.7"
sse-starlette = "^2.1.2"
psycopg2-binary = "^2.9.9"
lxml = "^5.2.2"
black = "^24.4.2"
twilio = "^9.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
